Merged PR 406: Refactoring towards supporting tasks
This PR includes various refactoring that I did in the P# `MachineTasks` branch towards supporting tasks and also generalizing and simplifying some stuff (as well as nuking some unnecessary stuff). It also moves a few stuff around.

I need to open a separate PR for some incoming changes to machine-related namespaces (so that tasks and machines are equal citizens of Coyote). And will also open a separate PR for the actual task implementation, which will be more self-contained after this PR.

Related work items: #699