Merged PR 506: Refactored random generator on its own static class
Immad gave feedback earlier that the random generator (boolean and integer) APIs should not be part of the `Specification` class. This feedback makes sense as these APIs are indeed orthogonal to the notion of a "specification", and they can be used both in production and mock/test code. For this reason, this PR refactors this bit of logic, and puts them in their own `RandomValueGenerator` class, which exposes these API as static methods.

This also addresses a previous work item, by removing the need to recreate a new `Random` instance each time the corresponding production runtime API is called (this API simply uses the random instance of the generator, created only once or whenever the seed is updated -- which is not thread safe as noted in the comment, if the user chooses to do that, its currently expected to only be done in the very beginning of the application).

This also moves towards simplifying the API surface of state machines, by deprecating the previous Random and RandomInteger APIs of machine, and now asking the user to use this new static methods.

Also changes all tests to this API (we need to change samples too after we finalize and merge).

@<Chris Lovett> @<Akash Lal>  please have a look and let me know if you have any suggestions.

Related work items: #758