Merged PR 590: Fixing nondet issue with halt
This hopefully fixes the nondet issue (causing flaky tests) with halting. There is a very subtle "race condition" (during the serialized testing execution) that caused a "window" where a machine has halted but not yet removed from the set of machines, and an enqueue happening, still enqueuing the event, but not triggering the "halted machine dropped event" check.