Merged PR 548: Initial porting of docs from psharp
This PR moves over the documentation from P# to Coyote, removes any stuff that are not part of Coyote (e.g. syntax), renamed P# mentions to Coyote. I also did a minor initial pass improving some of the text, but much more need to be done.

I don't plan this PR to be the polished version, rather than merge it quickly so that we have a documentation basis, and then we start iterating on smaller, focused, subsequent PRs.

Related work items: #659, #730, #918