Merged PR 674: Updating version to v1.0.0-rc1 and remove duplication
This PR changes the `internal.N` suffix to `rcN` (release candidate), so that we can start releasing the NuGet on nuget.org (towards the Azure customer release, although the nuget will be public, but not really advertised and thus discoverable yet). We can remove the `rcN` suffix once we have fully stabilized any breaking changes, and we are ready for a fully-public `1.0.0`. Once this is merged, I will update the Westlake PR to take a dependency on the release candidate from the public feed.

This PR also importantly removes duplication: there is no need any more to specify the version twice. The nuget creation script, will read the version from the corresponding property file inside `Commons`, and update the nuget accordingly. The `$version` inside the nuspec file is actually overridden by the `nuget pack` command line args, it just requires a dummy value to be given.

Also the latest nuget requires to specify framework dependencies, which are now added.

Also added the standard code of conduct mention in the README, and simplified the README for now, since all info is in our website.

@<Akash Lal>, @<Chris Lovett>

Related work items: #781