Merged PR 726: Fix in publish nuget script, add nuget icon, increment to rc2
This PR fixes the nuget source url in the nuget publish script (note that we do not use this script in CI, but good I guess to keep it in the repo, in case we ever need to publish manually from cmd line).

Also trying to add an icon in the nuget, and bumping up the rc version to enable the icon :) (can't test the icon until its publish to nuget.org, but followed precisely the instructions so hopefully should work...).