Merged PR 755: Renaming machine id to actor id
This PR renames MachineId (in the code, as well as some comments and in the docs) to ActorId **only**. It should not touch any of the stuff in Chris Machine -> StateMachine PR, so the two should be able to be merged in parallel.

More refactoring in separate PRs.

Related work items: #1157