Merged PR 790: fix bugs in the graph to do with monitors and halt.
The Graph writer was getting confused about "sender state" and monitor state so the monitor states were showing up in the sender.  There was also missing information during halt on what the current state was before the machine was halted, this missing information is gathered when we pop unhandled events, since Halt is always an unhandled event.