Merged PR 932: Typed state machine transitions
This PR changes the `StateMachine` API to make all transitions (`RaiseEvent`, `GotoState`, `PushState`, `PopState` and `Halt`) typed (via the new `StateMachine.Transition` type) towards improving the programmability of this programming model. The same applies to the `Monitor` API. Quite a bit of runtime complexity was also reduced based on the fact that the C# type checker does some of the work for us now.

This basically allows a `StateMachine` to be declared as follows:

```c#
private class M : StateMachine
{
  [Start]
  [OnEntry(nameof(InitOnEntry))]
  private class Init : State
  {
  }

  private Transition InitOnEntry() => this.GotoState<Done>();

  private class Done : State
  {
  }
}
```

Related work items: #1439