Merged PR 982: Fix bug in handling of raised events so the links show up correctly.
Fix bug in handling of raised events so the links show up correctly.  The problem was OnHandleRaisedEvent was trying to create a link immediately, but we don't yet know the impact of this raised event.  For example, it might result in a OnEventGotoState transition, in which case we want the link on the diagram to reflect that.  Fortunately the fix was very simple since I had a concept of dequeued events already.

Also added animating raft demo under http://127.0.0.1:4000/coyote/learn/core/demo.

Also added the beginnings of some tutorial content - more of a template really...