Merged PR 1047: Adding missing apis and fixing interleavings in controlled tasks
This PR adds some missing APIs from `ControlledTask`, namely the `Wait` methods.

The PR also addresses a super sneaky bug in the controlled task implementation (and also adds a few new tests to guard against regressions).

A simple repro:

```c#
private class SharedEntry
{
  public int Value = 0;
}

private static async ControlledTask WriteAsync(SharedEntry entry, int value)
{
  await ControlledTask.Delay(10);
  entry.Value = value;
}

private static async ControlledTask InvokeWriteAsync(SharedEntry entry, int value)
{
  await WriteAsync(entry, value);
}

[Test]
public static async ControlledTask RunExample()
{
  SharedEntry entry = new SharedEntry();
  ControlledTask task1 = InvokeWriteAsync(entry, 3);
  ControlledTask task2 = InvokeWriteAsync(entry, 5);
  await ControlledTask.WhenAll(task1, task2);
  Specification.Assert(entry.Value == 5, "Value is {0} instead of 5.", entry.Value);
}
```

The bug causes the assertion to never fail. Same behavior happens with some other combinations e.g. `Yield` instead of `Delay`.

The issue is that in the intermediate (`InvokeWriteAsync`) method (could be more than 1 of them) the "rewriting" that coyote does during systematic testing does not "package" the “after-await-logic” into a continuation and then return to the caller (`RunExample` here). Instead it "blocks" the current task executing until the await completes, which is wrong semantics, and causes the missing interleaving. Although here the assertion is missed, it could potentially lead to a worse scenario, e.g. a deadlock not getting resolved (imagine if `WriteAsync(3)` was waiting on a lock that only `WriteAsync(5)` releases 😊).

The pre-fix logic only did the packaging (which generates a special task under the hood, which is controlled and explored systematically) when a `ControlledTask` was created explicitly (`Run`, `Yield`, `Delay`) but not implicitly as in the above intermediate example, thus causing the issue! The actual logic (and production .NET semantics) is a bit more complicated than my description.

The fix is quite simple conceptually, imagine an async callstack like this:

```
WriteAsync
InvokeWriteAsync
RunExample
```

Coyote should do the packaging when an await happens in any method (which can be anonymous lambda), but not on the “root” method (`RunExample`, in this case). If it happens in the root, then it can hang the scheduler (as there is no caller method to return too). My current fix is to do a limited traversal of the `StackTrace` and differentiate between the cases.

The PR also does some minor related cleanup.