Merged PR 1088: Add Transition.None public static member because it is more discoverable than "default".
Add Transition.None and hide the Transition.Type since there is no need to make that stuff public, in fact it just leads to confusion because one starts to hunt for ways to change that field but there is no way.  Also added more documentation to Transition class and how to use it properly.

Related work items: #1946