Merged PR 1123: Fix bugs in coverage graph so that coverage report is correct.
Graph is consolidating events from one state to another so it shows up as one link, but if those are actually different event types it was forgetting that.  This fix changes the Graph data structure so it can remember multiple event id's on a single link.  The coverage report then uses that to produce the right result.  The handling of Monitors was also slightly wrong where the [hot], [cold] state labels were not matching with the coverage report MachineStates info, so this has been changed to a graph label only, rather than being part of the node id.

Also changed things so that you can now do this on the coyote test command line:
```
--explore --coverage activity
```

and you get the combined coverage of all explorations, regardless of whether any bugs were found.

Related work items: #2037