Merged PR 1158: Reducing array allocations in the testing runtime and cleanup
This PR reduces some array allocations in the testing runtime. Its also nukes some experimental strategies that we never really used (not even for paper evaluations) to cleanup the codebase. These strategies are in the P# codebase, so if we ever really need to bring them back and polish them, we can easily do them. I would normally clean them up in a separate PR, but doing it here as I had to change an API so they could not compile anymore due to a type difference.

This change reduced the peak working set quite a lot.