Merged PR 1172: Removing leftover ignore and defer from actor type
This PR removes the `IgnoreEvent` and `DeverEvent` APIs from the `Actor` type. I was playing with them when originally created the `Actor` type but forgot to remove them when we merged.

Basically, these APIs should not exist on an actor (similar to how actors in other frameworks/languages do not have them), as ignoring and deferring events is a state machine concept (in the sense that an event type can be ignored or deferred on some state, and then restored on another, but base actors do not have states). If someone really wanted (say to ignore an event because the actor doesn't care about it) they could just add that in the control flow of an event handler (e.g. if type of dequeued event is foo, then return).

Another bug that this PR fixes, is that because these APIs where on the base `Actor` type, where also exposed on the `StateMachine` type which is confusing and wrong (as they cannot alter the statically typed ignoring/deferring of state machines), so they were no-ops on the subclass.