Merged PR 1185: Refactoring to reuse random generator in runtime
This PR reuses the same `IRandomValueGenerator` that systematic testing strategies use in the runtime (instead of calling `Random` directly), so things are consistent (and in the future we easily plugin different random -- not that we currently need to do that, but you never know).

Also changes the name of the public static `RandomValueGenerator` API (used by `ControlledTask` objects) to `ControlledRandomValueGenerator`, first to match the naming of `ControlledTask` APIs and second to make it clear that this is a random generator that is controlled during testing.

Also adding an internal helper for creating a new instance of the production runtime (inside the existing runtime factory class), so that we can set the random value generator in as few places as possible.

**Note:** I am not trying to reinvent or improve the `IRandomValueGenerator` interface in this PR, rather just make sure we reuse what we currently have for consistency. Any improvements to the interface are outside the scope of this PR.