Merged PR 1290: Fix bugs related to having Actors in a state machine diagram.
Fix bugs related to having Actors in a state machine diagram.  Specifically the "senderState" on events is null and so we need to handle that properly, both in the DGML generation and the SVG animations.  Added animation to CoffeeMachine tutorial to test it out, since CoffeeMachine MockSensors is an Actor.