Merged PR 1434: refactoring to simplify testing engine
This PR does a simplification in the testing engine by encapsulating the logic of loading the information regarding a test method inside a `TestMethodInfo`, which is then used by the engine (instead of having that logic thrown inside the engine).

This is one step towards simplifying the engine logic in subsequent PRs.

This PR requires to first merge !1429 and !1430.