Merged PR 1417: added lock built on resources
This PR introduces the internal concept of a `Resource`, which asynchronous operation can await to acquire or release.

Built on top of that, the PR adds `AsyncLock` and `Semaphore`, two lock types that can be controlled during testing.

The PR also does some related refactoring to simplify the namespaces and reduce coupling between previous namespaces.

Related work items: #2649, #698