Merged PR 1576: update version to 1.0.1
This PR fixes an issue in the runtime and updates the version to 1.0.1.

Basically, there should always be a default task runtime instance (for cases were the user does not create a task runtime in production). The `Current` runtime should make sure to check if the runtime is controlled or not, as the behavior changes in testing. Also removed dead code in the runtime factories, fixed a remark and protecting the assignment of the installed task runtime using a lock.

Related work items: #2815, #2822