Merged PR 1646: cleanup in some internal types
Removed some code that is not required.

Removed the internal `ITaskController` interface, that was required when we originally had 2 projects, but now we don't need that and it also adds unnecessary perf cost going via the interface route (had to remove a few unused parameters in 3 internal APIs, will bring them back when we support them).

Removed `InterceptingTaskScheduler` as that's not needed, new logic that was added a while ago takes care of what this scheduler was trying to do in the past.