Merged PR 1612: Documentation for Semaphore in ... docs\learn\programming-models\async
This document is similar to the documentation for `TaskCompletionSource` and for `AsyncLock`.

It provides a summary, then explains why a `Semaphore` class is necessary in Coyote, then walks the reader through the available methods/property of this class.

Finally it presents a code sample with two scenarios -- one normal and one ending with a deadlock.