Merged PR 1917: There is a way to process inheritdocs during the build which means our nuget
There is a way to process inheritdocs during the build which means our nuget package will work properly with VS Code.  Before this fix, VS code shows no intellisense for those methods that have <inheritdocs/> tag in the shipped version of Microsoft.Coyote.xml.  With this fix our build takes care of it so that the nuget package version of Microsoft.Coyote.xml is fully resolved.

Related work items: #3350