Merged PR 1982: Fix code coverage and add support for code coverage on .net core 3.1
Fix code coverage and add support for code coverage on .net core 3.1.  Remove funky code that was trying to keep up to 10 sets of instrumented assemblies, renaming folders and all that jazz.  It is complicated, confusing and not needed.  Change how binaries are instrumented so the instrumented binaries are not copied in place over the one being tested, instead it is copied to the Coyote OutputPath and then the out of proc test is directed to load it from there.  Fix bug in how dependent assemblies are instrumented so that they are all instrumented regardless of whether the include Coyote or not.  A program might have a Utilities shared library that has no dependency on Coyote, and there's no reason not to instrument it as part of the test run.  In fact, we should instrument everything except Coyote itself and any mscorlib system libraries.

Related work items: #2826, #2827, #715