Merged PR 2245: synchronize accesses to the operation scheduler
Synchronizing access to API calls to the operation scheduler. This is part of refactoring to simplify the scheduler logic and bring it to the new scheduling design that we use in our language-agnostic [scheduler API](https://github.com/microsoft/coyote-scheduler/blob/master/include/coyote/scheduler.h).

This does not slow down the scheduler btw, e.g. on our CI previous and new tests terminate pretty much at the same time. This is because the scheduler was already serialized by how it was accessed by the runtime, so there is no lock contention. But this makes it easy when we cannot guarantee this serialization (e.g. if users create their own mocks, or we do various tricks with intercepting tasks as I am trying to do on my other WIP).

Related work items: #3920