Merged PR 2258: make bookkeeping of current op more robust
This PR simplifies the logic of bookkeeping the currently executing operation in `OperationScheduler` and make it more robust.

We remove reliance on a dictionary of task ids to operations, which not only simplifies the logic, but also makes it more robust because we cannot rely on task ids. This is because the .NET runtime does not guarantee that task ids will be unique. From the official [docs](https://docs.microsoft.com/en-us/dotnet/api/system.threading.tasks.task.id?view=netcore-3.1):

> Task IDs are assigned on-demand and do not necessarily represent the order in which task instances are created. Note that although collisions are very rare, task identifiers are not guaranteed to be unique.