Merged PR 2326: fix dispose on CoyoteRuntime and ControlledRuntime
Fix dispose on CoyoteRuntime so it resets the Current task and fix dispose on ControlledRuntime so it resets IsExecutionControlled. Add test to make sure it works.

This allows mixing of production tests in with systematic tests in the same test run which is important onramp for customers adopting Microsoft.Coyote.Task where they want to be able to run new Coyote tests and their old traditional unit tests in the same test run.

Related work items: #4065