Merged PR 2358: make run-tests script logger optional
We don't always need to dump `trx` files, I found that it was dumping tons and tons of them on my laptop from repeated runs, so I made it optional.

You can now pass a `-logger FORMAT` option to the script, where `FORMAT` is a supported format like `trx`, like this:
```
.\Scripts\run-tests.ps1 -logger trx
```

And this will dump files like this:
```
Results File: ...\coyote\Tests\Production.Tests\TestResults\pdeligia_2020-06-10_07_14_18.trx
```

If we ever need this in CI, we can enable it as an option. But I noticed that this does not really do much, besides dumping the results on disk. Perhaps DevOps has to be given the `TestResults` directory?