Merged PR 2392: optimize exploration with task delays using scheduler yields
Capability to specify to the scheduler that the program wants to yield the execution of the currently executing operation. This allows strategies to implement different techniques to deal with this. In this PR we implement de-prioritization with PCT. This is very useful for dealing with infinite (polling) task loops and (periodic) timers, but this PR only changes `Task.Yield`, we will deal with timers in a later PR.

In customer code, this reduced the exploration steps to reach cold state in a liveness monitor (and thus the end of a infinite test) from 10-20k down to 1k steps.

Related work items: #4218