Merged PR 3083: throw assertion exception on bug for MSTest rewriting
Our MSTest rewriting does not check if an iteration was finding a bug or not, which means a test would pass even if a bug was found. It only checked for unhandled exceptions but that is just a subset of all bugs that can be found. This is already discovering issues in some of the buggy Torch tests.