Update charts.py
source should be initialized as a ColumnDataSource object, instead of a dictionary. This is because Bokeh expects data to be provided in this format for efficient rendering.

methods(obj) seems to be a standalone function that is not called anywhere in the code. If this function is not needed, it can be removed.